requirements.txt
azure-functions==1.10.0
google-genai==0.1.0  # 정확한 버전은 설치 시 확인 후 지정
